`@remotion/studio`: Make render rows draggable
